requirements.txt
python-telegram-bot==20.3
httpx==0.24.0
beautifulsoup4==4.12.2
requests==2.31.0
supabase==1.1.0
